Creacion de PDFController, RegistroController. Ya puede generar PDF a partir de vistas, no puede realizar inserts en BD y falta implementar la prevista.
