--- CBTIS223_v2/Controllers/XXXXXXXXXXX.cs
@@ -1,12 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace CBTIS223_v2.Controllers
-{
-    public class XXXXXXXXXXX : Controller
-    {
-        public IActionResult Index()
-        {
-            return View();
-        }
-    }
-}
