+++ CBTIS223_v2/Views/Shared/Error.cshtml
@@ -1,25 +1,34 @@
-﻿@model ErrorViewModel
+﻿@using Microsoft.AspNetCore.Identity
 @{
-    ViewData["Title"] = "Error";
+    Layout = null;
 }
 
-<h1 class="text-danger">Error.</h1>
-<h2 class="text-danger">An error occurred while processing your request.</h2>
+<!DOCTYPE html>
+<html>
 
-@if (Model.ShowRequestId)
-{
-    <p>
-        <strong>Request ID:</strong> <code>@Model.RequestId</code>
-    </p>
-}
+<head>
+    <title>Documentos</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
+          integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC"
+          crossorigin="anonymous">
 
-<h3>Development Mode</h3>
-<p>
-    Swapping to <strong>Development</strong> environment will display more detailed information about the error that occurred.
-</p>
-<p>
-    <strong>The Development environment shouldn't be enabled for deployed applications.</strong>
-    It can result in displaying sensitive information from exceptions to end users.
-    For local debugging, enable the <strong>Development</strong> environment by setting the <strong>ASPNETCORE_ENVIRONMENT</strong> environment variable to <strong>Development</strong>
-    and restarting the app.
-</p>
+    <link rel="stylesheet" type="text/css" href="~/css/documentos.css" />
+
+    <script src="https://code.jquery.com/jquery-3.6.0.min.js"
+            integrity="sha384-KyZXEAg3QhqLMpG8r+PW8l5O8R18EJSf5PPSb5l5p5zqo4pE2F5/18z5p5z5E5Q5v5z5g5l5p5z5E5"
+            crossorigin="anonymous"></script>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js"
+            integrity="sha384-k4GT4mMEkCk4OfF5/MF5Lm5wKj5l5T5u5v5C5n5u5O5x5V5B5r5Z5E5O5W5B5V5E5T5u5C5"
+            crossorigin="anonymous"></script>
+</head>
+<body class="bg-white">
+    <h1>Error</h1>
+
+    <div class="alert alert-danger">
+        <p>Error al procesar tu solicitud</p>
+        <a asp-action="Documentos" asp-controller="Home" class="btn btn-primary">Volver</a>
+    </div>
+</body>
+
+</html>
